+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Dependencies
 - `typescript-eslint`: 8.24.1 -> 8.25.0
 - `@stylistic/eslint-plugin`: 4.0.1 -> 4.1.0
+- `@vitest/eslint-plugin`: 1.1.31 -> 1.1.36
 
 ## 0.5.0